+++ src/ContextProvider/index.js
@@ -5,28 +5,28 @@ import { Component, PropTypes, Children } from "react"
 export default class StatinamicContext extends Component {
   props: {
     collection: StatinamicCollection,
-    layouts: Object,
+    layouts: Object, // deprecated
     metadata: Object,
     children?: any,
   };
 
   static propTypes = {
     collection: PropTypes.array,
-    layouts: PropTypes.object,
+    layouts: PropTypes.object, // deprecated
     metadata: PropTypes.object,
     children: PropTypes.oneOfType([ PropTypes.array, PropTypes.object ]),
   };
 
   static childContextTypes = {
     collection: PropTypes.array,
-    layouts: PropTypes.object,
+    layouts: PropTypes.object, // deprecated
     metadata: PropTypes.object,
   };
 
   getChildContext(): Object {
     return {
       collection: this.props.collection,
-      layouts: this.props.layouts,
+      layouts: this.props.layouts, // deprecated
       metadata: this.props.metadata,
     }
   }