+++ src/PageContainer/__tests__/component.js
@@ -16,12 +16,148 @@ const Page = () => <div className="Page" />
 const PageError = () => <div className="PageError" />
 const AnotherPage = () => <div className="AnotherPage" />
 
-// Don't print noisy log unless I mocked you
-test.beforeEach(() => {
-  console.info = noop
+test("should render a Page if page is ok", () => {
+  const renderer = createRenderer()
+  renderer.render(
+    jsx(
+      PageContainer,
+      {
+        params: { splat: "" },
+        pages: { "/": {} },
+        getPage: noop,
+        setPageNotFound: noop,
+        layouts: { Page },
+      }
+    ),
+    {
+      collection: [],
+    },
+  )
+  expect(renderer.getRenderOutput()).toEqualJSX(
+    <div>
+      <Page ref={ function noRefCheck() {} } />
+    </div>
+  )
 })
 
-test("should render a Page if page is ok", () => {
+test.cb("should try to get a page if no page in cache", (t) => {
+  const renderer = createRenderer()
+  renderer.render(
+    jsx(
+      PageContainer,
+      {
+        params: { splat: "" },
+        pages: { },
+        getPage: (pageUrl, dataUrl) => {
+          t.is(pageUrl, "/")
+          t.is(dataUrl, "/j.son")
+          t.end()
+        },
+        setPageNotFound: () => {
+          t.fail()
+          t.end()
+        },
+        layouts: { Page },
+      }
+    ),
+    {
+      collection: [
+        {
+          __url: "/",
+          __dataUrl: "/j.son",
+        },
+      ],
+    },
+  )
+  renderer.getRenderOutput()
+})
+
+test(`should render a visible error if page is not ok and no PageError
+available`, () => {
+  const renderer = createRenderer()
+  renderer.render(
+    jsx(
+      PageContainer,
+      {
+        params: { splat: "" },
+        pages: { "/": { error: "Test", errorText: "" } },
+        getPage: noop,
+        setPageNotFound: noop,
+        layouts: { Page },
+      }
+    ),
+    {
+      collection: [],
+    },
+  )
+
+  expect(renderer.getRenderOutput()).toEqualJSX(
+    <div>
+      <div style={ { "text-align": "center" } }>
+        <h1>
+          { "Test" }
+        </h1>
+        <p />
+      </div>
+    </div>
+  )
+})
+
+test("should render a PageError if page is not ok and PageError is available",
+() => {
+  const renderer = createRenderer()
+  renderer.render(
+    jsx(
+      PageContainer,
+      {
+        params: { splat: "" },
+        pages: { "/": { error: "Test" } },
+        getPage: noop,
+        setPageNotFound: noop,
+        layouts: { Page, PageError },
+      }
+    ),
+    {
+      collection: [],
+    },
+  )
+
+  expect(renderer.getRenderOutput()).toEqualJSX(
+    <div>
+      <PageError error="Test" />
+    </div>
+  )
+})
+
+test("should render a another page layout if defaultLayout is used", () => {
+  const renderer = createRenderer()
+  renderer.render(
+    jsx(
+      PageContainer,
+      {
+        params: { splat: "" },
+        pages: { "/": {} },
+        getPage: noop,
+        setPageNotFound: noop,
+        defaultLayout: "AnotherPage",
+        layouts: { AnotherPage },
+      }
+    ),
+    {
+      collection: [],
+    },
+  )
+
+  expect(renderer.getRenderOutput()).toEqualJSX(
+    <div>
+      <AnotherPage ref={ function noRefCheck() {} } />
+    </div>
+  )
+})
+
+// deprecated
+// until the end of the file :)
+test("should render a Page if page is ok (context)", () => {
   const renderer = createRenderer()
   renderer.render(
     jsx(
@@ -45,7 +181,7 @@ test("should render a Page if page is ok", () => {
   )
 })
 
-test.cb("should try to get a page if no page in cache", (t) => {
+test.cb("should try to get a page if no page in cache (context)", (t) => {
   const renderer = createRenderer()
   renderer.render(
     jsx(
@@ -78,7 +214,7 @@ test.cb("should try to get a page if no page in cache", (t) => {
 })
 
 test(`should render a visible error if page is not ok and no PageError
-available`, () => {
+available (context)`, () => {
   const renderer = createRenderer()
   renderer.render(
     jsx(
@@ -108,7 +244,8 @@ available`, () => {
   )
 })
 
-test("should render a PageError if page is not ok and PageError is available",
+test(`should render a PageError if page is not ok and PageError is available
+(context)`,
 () => {
   const renderer = createRenderer()
   renderer.render(
@@ -134,7 +271,8 @@ test("should render a PageError if page is not ok and PageError is available",
   )
 })
 
-test("should render a another page layout if defaultLayout is used", () => {
+test("should render a another page layout if defaultLayout is used (context)",
+() => {
   const renderer = createRenderer()
   renderer.render(
     jsx(