+++ src/PageContainer/component.js
@@ -2,7 +2,29 @@
 /* eslint-disable react/sort-comp */
 import React, { Component, PropTypes } from "react"
 import { findDOMNode } from "react-dom"
+
 import urlify from "../_utils/urlify"
+
+type DefaultProps = {
+  defaultLayout: string,
+}
+
+type Props = {
+  pages: Object,
+  params: {
+    splat: string,
+  },
+  layouts: Object,
+  defaultLayout: string,
+  getPage: Function,
+  setPageNotFound: Function,
+  logger: Object,
+}
+
+type Context = {
+  collection: StatinamicCollection,
+  layouts: Object // deprecated
+}
 
 // react-router does not return leading and trailing slashes
 // so we need to normalize according to collection data
@@ -31,46 +53,81 @@ function find(
   ))
 }
 
-type DefaultProps = {
-  defaultLayout: string,
-}
-type Props = {
-  pages: Object,
-  params: {
-    splat: string
-  },
-  defaultLayout: string,
-  getPage: Function,
-  setPageNotFound: Function,
+function getBase(location: Object): string {
+  return (
+    location.protocol + "//" + location.host + process.env.STATINAMIC_PATHNAME
+  )
 }
 
-type Context = {
-  collection: StatinamicCollection,
-  layouts: Object
+function adjustCurrentUrl(location: Object, item: Object, props: Props): void {
+  // adjust url (eg: missing trailing slash)
+  const currentExactPageUrl = location.href
+    .replace(getBase(location), "/")
+    .replace(location.hash, "")
+
+  if (currentExactPageUrl !== item.__url) {
+    props.logger.info(
+      "statinamic: PageContainer: " +
+      `replacing by '${ currentExactPageUrl }' to '${ item.__url }'`
+    )
+    if (browserHistory) {
+      browserHistory.replace(item.__url)
+    }
+  }
+}
+
+let layoutFromContextWarning = false
+function getLayout(
+  layout: string, props: Props, context: Context, warn: boolean = true
+): ReactClass | void {
+  if (props.layouts && props.layouts[layout]) {
+    return props.layouts[layout]
+  }
+
+  if (context.layouts && context.layouts[layout]) {
+    if (warn && !layoutFromContextWarning) {
+      props.logger.warn(
+        "statinamic: You are using a layout defined in the client and build  " +
+        `scripts ('${ layout }'). \n` +
+        "This method is deprecated and will be removed in the future. \n" +
+        "In order to have more flexibility, you should create your own " +
+        "PageContainer and provide layouts to it via a `layouts` prop. " +
+        "This will allow your to have more control over components by being "+
+        "more explicit. \n"+
+        "Check out migration instruction in the CHANGELOG. "
+      )
+      layoutFromContextWarning = true
+    }
+    return context.layouts[layout]
+  }
 }
 
 class PageContainer extends Component<DefaultProps, Props, void> {
   _content: Element;
 
-  static propTypes = {
-    pages: PropTypes.object.isRequired,
-    params: PropTypes.object,
-
-    defaultLayout: PropTypes.string,
-
-    // actions
-    getPage: PropTypes.func.isRequired,
-    setPageNotFound: PropTypes.func.isRequired,
-  };
+  propTypes: Props;
 
   static contextTypes = {
-    collection: PropTypes.array.isRequired,
-    layouts: PropTypes.object.isRequired,
+    collection: PropTypes.arrayOf(PropTypes.object),
+    layouts: PropTypes.object,
   };
 
-  static defaultProps = {
+  static defaultProps: DefaultProps = {
+    layouts: {},
     defaultLayout: "Page",
+    logger: console,
   };
+
+  constructor(props: Props, context: Context) {
+    super(props)
+
+    if (!getLayout(props.defaultLayout, props, context)) {
+      props.logger.error(
+        "statinamic: PageContainer: " +
+        `default layout "${ props.defaultLayout }" not provided. `
+      )
+    }
+  }
 
   componentWillMount() {
     this.preparePage(this.props, this.context)
@@ -111,44 +168,16 @@ class PageContainer extends Component<DefaultProps, Props, void> {
   }
 
   preparePage(props: Props, context: Context): void {
-    if (!context.layouts[props.defaultLayout]) {
-      console.error(
-        "statinamic: PageContainer: " +
-        `default layout "${ props.defaultLayout }" doesn't exist. ` +
-        "Please check your configuration (\"layouts\" part). " +
-        `If you haven't defined "${ props.defaultLayout }", you should. `
-      )
-    }
-
     const pageUrl = splatToUrl(props.params.splat)
     if (isDevelopmentClient()) {
-      console.info(`statinamic: PageContainer: '${ pageUrl }' rendering...`)
+      props.logger.info(
+        `statinamic: PageContainer: '${ pageUrl }' rendering...`
+      )
     }
 
     const item = find(context.collection, pageUrl)
     if (isClient() && item) {
-      // adjust url (eg: missing trailing slash)
-      const currentExactPageUrl = window.location.href
-        .replace(
-          (
-            window.location.protocol +
-            "//" +
-            window.location.host +
-            process.env.STATINAMIC_PATHNAME
-          ),
-          "/"
-        )
-        .replace(window.location.hash, "")
-
-      if (currentExactPageUrl !== item.__url) {
-        console.info(
-          "statinamic: PageContainer: " +
-          `replacing by '${ currentExactPageUrl }' to '${ item.__url }'`
-        )
-        if (browserHistory) {
-          browserHistory.replace(item.__url)
-        }
-      }
+      adjustCurrentUrl(window.location, item, props)
     }
 
     const page = props.pages[pageUrl]
@@ -157,9 +186,8 @@ class PageContainer extends Component<DefaultProps, Props, void> {
         props.getPage(item.__url, item.__dataUrl)
       }
       else {
-        console.error(
-          "statinamic: PageContainer: " +
-          `${ pageUrl } is a page not found.`
+        props.logger.error(
+          `statinamic: PageContainer: ${ pageUrl } is a page not found.`
         )
         props.setPageNotFound(pageUrl)
       }
@@ -169,9 +197,9 @@ class PageContainer extends Component<DefaultProps, Props, void> {
         return
       }
 
-      const Layout = this.getLayout(props, context, page)
+      const Layout = getLayout(page.type, props, context)
       if (page.type !== undefined && !Layout) {
-        console.error(
+        props.logger.error(
           "statinamic: PageContainer: " +
           `Unkown page type: "${ page.type }" component not available in ` +
           "\"layouts\" property. " +
@@ -181,36 +209,36 @@ class PageContainer extends Component<DefaultProps, Props, void> {
     }
   }
 
-  getLayout(props: Props, context: Context, page: Object): ReactClass {
-    return context.layouts[page.type || props.defaultLayout]
-  }
-
   render() {
-    const pageUrl = splatToUrl(this.props.params.splat)
-    const page = this.props.pages[pageUrl]
+    const { props, context } = this
+
+    const pageUrl = splatToUrl(props.params.splat)
+    const page = props.pages[pageUrl]
 
     if (!page) {
       if (isDevelopmentClient()) {
-        console.info(`statinamic: PageContainer: '${ pageUrl }' no data`)
+        props.logger.info(`statinamic: PageContainer: '${ pageUrl }' no data`)
       }
       return null
     }
     if (isDevelopmentClient()) {
-      console.info(`statinamic: PageContainer: '${ pageUrl }'`, page)
+      props.logger.info(`statinamic: PageContainer: '${ pageUrl }'`, page)
     }
 
     if (
       typeof page !== "object" ||
       page.toString() !== "[object Object]"
     ) {
-      console.info(
+      props.logger.info(
         `statinamic: PageContainer: page ${ pageUrl } should be an object`
       )
       return null
     }
 
-    const { PageLoading, PageError } = this.context.layouts
-    const Layout = this.getLayout(this.props, this.context, page)
+    const PageLoading = getLayout("PageLoading", props, context, false)
+    const PageError = getLayout("PageError", props, context, false)
+    const LayoutFallback = getLayout(props.defaultLayout, props, context, false)
+    const Layout = getLayout(page.type, props, context, false) || LayoutFallback
 
     return (
       <div>